migrating heap script to its own module
- ignoring module from linting
