+++ .eslintignore
@@ -4,3 +4,5 @@
 .env
 .env.*
 !.env.example
+
+src/lib/heap.js
