+++ src/pages/_app.jsx
@@ -6,44 +6,7 @@ import { Layout } from "@/components/Layout";
 import "focus-visible";
 import "@/styles/tailwind.css";
 
-if (
-  typeof window !== "undefined" &&
-  window.location.host === "docs.goldsky.com"
-) {
-  (window.heap = window.heap || []),
-    (heap.load = function (e, t) {
-      (window.heap.appid = e), (window.heap.config = t = t || {});
-      var r = document.createElement("script");
-      (r.type = "text/javascript"),
-        (r.async = !0),
-        (r.src = "https://cdn.heapanalytics.com/js/heap-" + e + ".js");
-      var a = document.getElementsByTagName("script")[0];
-      a.parentNode.insertBefore(r, a);
-      for (
-        var n = function (e) {
-            return function () {
-              heap.push([e].concat(Array.prototype.slice.call(arguments, 0)));
-            };
-          },
-          p = [
-            "addEventProperties",
-            "addUserProperties",
-            "clearEventProperties",
-            "identify",
-            "resetIdentity",
-            "removeEventProperty",
-            "setEventProperties",
-            "track",
-            "unsetEventProperty",
-          ],
-          o = 0;
-        o < p.length;
-        o++
-      )
-        heap[p[o]] = n(p[o]);
-    });
-  heap.load("303099433");
-}
+import "../lib/heap";
 
 function getNodeText(node) {
   let text = "";
